move `waitFor` util to core/utils
